+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Improve conversion accuracy for CInt [#658](https://github.com/icsharpcode/CodeConverter/pull/658)
 * Create non-static delegate when converting a shared event[#671](https://github.com/icsharpcode/CodeConverter/issues/671)
 * Improve implicit conversion of delegates [#632](https://github.com/icsharpcode/CodeConverter/issues/632)
+* Retain enum names in nullable Select Case [#675](https://github.com/icsharpcode/CodeConverter/issues/675)
 
 ### C# -> VB
 * Generics type parameter missed BC30737 and BC32042 [#682] (https://github.com/icsharpcode/CodeConverter/issues/682)